Fix DeleterThread and CacheCleanerThread locking. (#877)
This change fixes double unlock that was causing crashes when
the Tests binary would exit on Windows.

This change replaces the explicit unlock() on the mutex with a
call to unlock() on the QMutexLocker holding the lock. This allows
to QMutexLocker to avoid calling unlock() in its destructor when
the mutex has already been unlocked.

The QMutexLocker that is declared right after the unlock() was also
given a unique name in the function scope. This has no functional
effect, but could help avoid accidentally calling the wrong object
if this code is modified in the future.