Fix CrashReporter process invocation on Windows. (#880)
This change fixes the CrashReporter code that launches the actual
Natron binary. It removes a parameter that was not needed on Windows
that caused the Natron binary to immediately exit. It also fixes
the command-line parameters passed to the child process. The old
code was passing in the CrashReporter binary name as the first
parameter which was causing the command-line parsing code to fail
because it interpreted this as an unknown/unhandled parameter.