Fix libsox MinGW package
Recent changes in MinGW caused the libssp package and DLL to
no longer exist. This change caused the libsox build to fail to
generate a shared library. Without the libsox DLL, the openfx-arena
plugin build fails. This change removes a line from the
configure script that tries to detect libssp. This fixes the
libsox package build so it builds static and shared libraries
again.